+++ tests/testthat/test-clone.R
@@ -1077,6 +1077,27 @@ test_that("Cloning inherited methods for non-portable classes", {
   expect_identical(a$x, 3)
 })
 
+test_that("In deep_clone(), don't try to clone non-R6 objects", {
+
+  `$.test` <- function(x, value) {
+    stop("error")
+  }
+
+  AC <- R6Class("AC",
+    public = list(
+      x = NULL,
+      initialize = function() {
+        x <- new.env(parent = emptyenv())
+        class(x) <- "test"
+        self$x <- x
+      }
+    )
+  )
+
+  obj <- AC$new()
+  obj2 <- obj$clone(deep = TRUE)
+  expect_identical(obj$x, obj2$x)
+})
 
 test_that("Deep cloning", {
   AC <- R6Class("AC", public = list(x = 1))